refactor(MJCore): Resolve circular dependencies and add input validation
This commit resolves circular dependency issues that were preventing proper
validation in the GetEntityObject method and improves the overall architecture
of the MJCore package.

Changes:
1. Added input validation to GetEntityObject in metadata.ts
   - Validates entityName is a non-empty string
   - Validates contextUser is either null/undefined or a UserInfo object
   - These validations previously caused runtime errors due to circular deps

2. Extracted authorization logic to new authEvaluator.ts
   - Moved EvaluateAuthorizationDynamic from UserInfo class
   - Breaks circular dependency between metadata.ts and securityInfo.ts
   - Improves separation of concerns

3. Extracted metadata utilities to new metadataUtil.ts
   - Moved GetDataFromMemory from util.ts
   - Prevents circular dependency with metadata.ts
   - Creates cleaner module boundaries

4. Updated exports in index.ts
   - Added exports for new authEvaluator and metadataUtil modules
   - Maintains backward compatibility

These changes improve code organization, eliminate circular dependencies,
and enable proper runtime validation without breaking the module loading.
